feat: 매칭예상

See merge request s11-bigdata-recom-sub2/S11P22B106!234

+++ backend-recommend/catchcatch-recommend/src/main/java/catchcatch/recommend/domain/recommend/requestdto/MatchingDataDto.java
@@ -5,7 +5,7 @@ import lombok.Builder;
 @Builder
 public record MatchingDataDto (
      Integer rating,
-     Long entryTime,
+     Integer entryTime,
      String dayOfWeek,
      Long matchDuration,
      Integer currentUserSize){
